requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==4.2.11
django-admin==2.0.2
django-cors-headers==4.3.1
django-environ==0.11.2
django-excel-response2==3.0.6
django-localflavor==4.0
django-six==1.0.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
excel-base==1.0.4
isoweek==1.3.3
psycopg2==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-stdnum==1.19
pytz==2024.1
screen==1.0.1
six==1.16.0
sqlparse==0.4.4
TimeConvert==3.0.13
typing_extensions==4.10.0
tzlocal==5.2
xlwt==1.3.0
